Bring order to contact list
Changes:
1. Move team emails to secret storage. For operations they are not safe. 
We will use them for emergancy in gerenral.

2. Remove https segments from github, telegram and keybase fields. More 
compact and usable for web apps as the same id's can be used across 
domains.

3. Remove bitcoin address. Our bitcoins left and we dont expect to used 
it for payout purposes in the future. Also bitcoin addresses has not 
been used by everybody, hence the decision to kill them.

4. Rename etherem field to payout. Motivation is the following: teammate 
can choose to use the same address for gitcoin and salary to collect 
more reputation on this address. But for somebody it can be more usable 
to have different addresses. Thus for every cricitical congress function 
we will have ethereum address. But they can be the same.